requirements.txt
# requirements.txt
apache-airflow-providers-postgres==6.2.3
# apache-airflow-providers-mysql==6.3.3
apache-airflow-providers-google==17.1.0
apache-airflow-providers-docker==4.4.2
apache-airflow-providers-trino==6.5.0
psycopg2-binary==2.9.10